FUSE-2520 Use InlineDialog to list team names when there are more than 3 teams. As part of this, added add-user-bitbucket-access-extension.soy.
Changed the way we check for number of teams with default groups. Was done as a condition and now is done as an if check inside velocity.

Changed the way we check whether we should checked the Bitbucket access checkbox on page load. Was done via server side check but decided we should just see whether JIRA Software access is checked (via JS code) to ensure consistency.

Modified AddUserBitbucketAccessExtensionContextProvider.java: removed redundant keys and call require data to provide data for the inline dialog.