FUSE-2520 Removed the hidden input from velocity. This should be something that gets dynamically added by our JS code. We should do this before form submission only if the Bitbucket checkbox is selected.
Removed inviteToGroups from the context provider. This was in the context provider so we can form the hidden input inside velocity. Since we're not doing this anymore, we don't need this in the context provider. However, we need to make this information available to our JS code. We do this via require data.